refactor: move Home into ArchivedComponents
