+++ src/components/ArchivedComponents/Home/HomeSiteCards.tsx
@@ -1,5 +1,4 @@
-import useAnalytics from "../../../context/Analytics/useAnalytics";
-
+import useAnalytics from "@/context/Analytics/useAnalytics";
 import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import CardLinkIcon from "@/components/SharedComponents/Card/CardLinkIcon";
