+++ src/components/Posts/PostList/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./PostList";
+export type { PostListProps } from "./PostList";
